Seperated Scrabble class into seperate file from Scrabble module file.
