+++ lib/scrabble/player.rb
@@ -1,0 +1,5 @@
+module Scrabble
+  class Player
+
+  end
+end
